+++ src/Samples/GameSessionViewModel.cpp
@@ -18,7 +18,7 @@ bool GameSessionViewModel::isInGameSession()
 	return gameSession->isInSession();
 }
 
-void GameSessionViewModel::setPeerReady()
+void GameSessionViewModel::setPlayerReady()
 {
 
 	auto client = Stormancer::IClientFactory::GetClient(this->parent->id);
@@ -42,6 +42,29 @@ void GameSessionViewModel::setPeerReady()
 	});
 }
 
+void GameSessionViewModel::leaveGameSession()
+{
+	auto client = Stormancer::IClientFactory::GetClient(this->parent->id);
+
+	auto gameSession = client->dependencyResolver().resolve<Stormancer::GameSessions::GameSession>();
+
+	this->parent->isProcessing = true;
+	gameSession->disconnectFromGameSession().then([this](pplx::task<void> t)
+	{
+
+		this->parent->isProcessing = false;
+		try
+		{
+			t.get();
+		}
+		catch (std::exception& ex)
+		{
+			this->parent->lastError = ex.what();
+		}
+
+	});
+}
+
 std::vector<P2PRemotePeerViewModel> GameSessionViewModel::getP2PRemotePeers()
 {
 	auto client = Stormancer::IClientFactory::GetClient(this->parent->id);
@@ -49,14 +72,17 @@ std::vector<P2PRemotePeerViewModel> GameSessionViewModel::getP2PRemotePeers()
 	auto gameSession = client->dependencyResolver().resolve<Stormancer::GameSessions::GameSession>();
 
 	std::vector<P2PRemotePeerViewModel> results;
-	for (auto p : gameSession->scene()->remotePeers())
+	auto peers = gameSession->scene()->connectedPeers();
+	for (auto p : peers)
 	{
 		P2PRemotePeerViewModel vm;
-		auto p2pPeer = std::static_pointer_cast<Stormancer::IP2PScenePeer>(p);
 
-		vm.isRelay = p2pPeer->useRelay();
-		vm.sessionId = p2pPeer->sessionId();
+
+
+		vm.isRelay = p.second->useRelay();
+		vm.sessionId = p.second->sessionId();
 		results.push_back(vm);
+
 	}
 	return results;
 }