+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.PartyMerging/PartyEventHandler.cs
@@ -1,8 +1,11 @@
-﻿using Stormancer.Server.Plugins.Party;
+﻿using Stormancer.Server.Plugins.Models;
+using Stormancer.Server.Plugins.Party;
+using Stormancer.Server.Plugins.Party.Dto;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Stormancer.Server.Plugins.PartyMerging
@@ -27,19 +30,42 @@ namespace Stormancer.Server.Plugins.PartyMerging
                 case PartyGameFinderStateChange.Stopped:
                     _service.TryRestartMerging();
                     break;
-               
+
                 case PartyGameFinderStateChange.Started:
                     try
                     {
                         await _service.CancelMerging();
                     }
-                    catch(OperationCanceledException)
+                    catch (OperationCanceledException)
                     {
 
                     }
                     break;
- 
+
             }
+        }
+
+        Task IPartyEventHandler.OnCreatingReservation(Stormancer.Server.Plugins.Party.CreateReservationContext ctx)
+        {
+            if (ctx.Reservation.CustomData.TryGetValue("merger", out var mergerId))
+            {
+                _ = ctx.Party.UpdateSettings(state =>
+                {
+
+
+                    var partySettings = new PartySettingsDto(state);
+                    if (partySettings.PublicServerData == null)
+                    {
+                        partySettings.PublicServerData = new Dictionary<string, string>();
+                    }
+                    partySettings.PublicServerData["stormancer.partyMerging.merged"] = "true";
+                    return partySettings;
+
+
+                }, CancellationToken.None);
+            }
+
+            return Task.CompletedTask;
         }
     }
 }