+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -34,6 +34,7 @@ Changed
 - If Configuration.HostSessionId exists but is empty, the game session now takes the first connected peer as host.
 - Try to connect agents in parallel.
 - Don't throw an error when trying to keep alive or stop a game server and the agent is offline.
+- Remove verbose Info logs.
 
 Fixed
 *****